refactor(ui): split CommunityBrowser into sibling sections
Extract TapChip, CompatibilityBadge, CommunityTapManagementSection, and
CommunityProfilesSection into src/components/community/. Parent CommunityBrowser
composes the two sibling sections plus cached-fallback banner and import modal;
error state and useMemo(visibleEntries) stay in the parent per task spec.
