+++ src/crosshook-native/src/components/CommunityBrowser.tsx
@@ -6,13 +6,13 @@ import {
   type CommunityProfileIndexEntry,
   type CommunityTapSubscription,
   type CommunityTapSyncResult,
-  deriveCommunityImportProfileName,
   type UseCommunityProfilesResult,
   useCommunityProfiles,
 } from '../hooks/useCommunityProfiles';
 import CommunityImportWizardModal from './CommunityImportWizardModal';
-import { CollapsibleSection } from './ui/CollapsibleSection';
-import { ThemedSelect } from './ui/ThemedSelect';
+import { CommunityProfilesSection } from './community/CommunityProfilesSection';
+import { CommunityTapManagementSection } from './community/CommunityTapManagementSection';
+import { ratingOrder } from './community/CompatibilityBadge';
 
 export interface CommunityBrowserProps {
   profilesDirectoryPath?: string;
@@ -20,16 +20,6 @@ export interface CommunityBrowserProps {
 }
 
 const DEFAULT_PROFILES_DIRECTORY = '~/.config/crosshook/profiles';
-
-const ratingOrder: CommunityCompatibilityRating[] = ['platinum', 'working', 'partial', 'broken', 'unknown'];
-
-const ratingLabel: Record<CommunityCompatibilityRating, string> = {
-  unknown: 'Unknown',
-  broken: 'Broken',
-  partial: 'Partial',
-  working: 'Working',
-  platinum: 'Platinum',
-};
 
 function matchesQuery(entry: CommunityProfileIndexEntry, query: string): boolean {
   const normalized = query.trim().toLowerCase();
@@ -85,77 +75,6 @@ async function chooseCommunityProfileImport(): Promise<string | null> {
   }
 
   return result ?? null;
-}
-
-function TapChip({
-  tap,
-  onRemove,
-  onPin,
-  onUnpin,
-  headCommit,
-  busy,
-}: {
-  tap: CommunityTapSubscription;
-  onRemove: (tap: CommunityTapSubscription) => void;
-  onPin: (tap: CommunityTapSubscription) => void;
-  onUnpin: (tap: CommunityTapSubscription) => void;
-  headCommit?: string;
-  busy: boolean;
-}) {
-  const shortPinnedCommit = tap.pinned_commit ? tap.pinned_commit.slice(0, 12) : null;
-  const shortHeadCommit = headCommit ? headCommit.slice(0, 12) : null;
-
-  return (
-    <div className="crosshook-community-tap">
-      <div className="crosshook-community-tap__meta">
-        <strong className="crosshook-community-tap__url">{tap.url}</strong>
-        <span className="crosshook-community-tap__branch">
-          {tap.branch ? `Branch: ${tap.branch}` : 'Default branch'}
-        </span>
-        <span className="crosshook-community-tap__branch">
-          {shortPinnedCommit ? `Pinned: ${shortPinnedCommit}` : `Tracking: ${shortHeadCommit ?? 'unsynced'}`}
-        </span>
-      </div>
-      <div className="crosshook-community-browser__button-row">
-        {tap.pinned_commit ? (
-          <button
-            type="button"
-            className="crosshook-button crosshook-button--secondary"
-            onClick={() => onUnpin(tap)}
-            disabled={busy}
-          >
-            Unpin
-          </button>
-        ) : (
-          <button
-            type="button"
-            className="crosshook-button crosshook-button--secondary"
-            onClick={() => onPin(tap)}
-            disabled={busy || !headCommit}
-            title={headCommit ? 'Pin this tap to the currently synced commit' : 'Sync taps first to capture a commit'}
-          >
-            Pin to Current Version
-          </button>
-        )}
-        <button
-          type="button"
-          className="crosshook-button crosshook-button--secondary"
-          onClick={() => onRemove(tap)}
-          disabled={busy}
-        >
-          Remove
-        </button>
-      </div>
-    </div>
-  );
-}
-
-function CompatibilityBadge({ rating }: { rating: CommunityCompatibilityRating }) {
-  return (
-    <span className={`crosshook-community-rating-badge crosshook-community-rating-badge--${rating}`}>
-      {ratingLabel[rating]}
-    </span>
-  );
 }
 
 export function CommunityBrowser({ profilesDirectoryPath = DEFAULT_PROFILES_DIRECTORY, state }: CommunityBrowserProps) {
@@ -281,260 +200,70 @@ export function CommunityBrowser({ profilesDirectoryPath = DEFAULT_PROFILES_DIRE
         </div>
       ) : null}
 
-      <CollapsibleSection title="Tap Management" className="crosshook-panel crosshook-community-browser__panel">
-        <div className="crosshook-community-browser__footer">
-          <div className="crosshook-community-browser__section-copy">
-            <p className="crosshook-muted crosshook-community-browser__helper">
-              Taps are persisted in CrossHook settings and synced through the backend community commands.
-            </p>
-          </div>
-          <div className="crosshook-community-browser__button-row">
-            <button
-              type="button"
-              className="crosshook-button crosshook-button--secondary"
-              onClick={() => {
-                void refreshProfiles().catch((refreshError) => {
-                  setError(refreshError instanceof Error ? refreshError.message : String(refreshError));
-                });
-              }}
-              disabled={loading || syncing}
-            >
-              Refresh Index
-            </button>
-            <button
-              type="button"
-              className="crosshook-button"
-              onClick={() => {
-                void syncTaps().catch((syncError) => {
-                  setError(syncError instanceof Error ? syncError.message : String(syncError));
-                });
-              }}
-              disabled={loading || syncing || taps.length === 0}
-            >
-              {syncing ? 'Syncing...' : 'Sync Taps'}
-            </button>
-          </div>
-        </div>
+      <CommunityTapManagementSection
+        taps={taps}
+        tapUrl={tapUrl}
+        tapBranch={tapBranch}
+        loading={loading}
+        syncing={syncing}
+        onTapUrlChange={setTapUrl}
+        onTapBranchChange={setTapBranch}
+        onAddTap={() => {
+          void handleAddTap();
+        }}
+        onRefresh={() => {
+          void refreshProfiles().catch((refreshError) => {
+            setError(refreshError instanceof Error ? refreshError.message : String(refreshError));
+          });
+        }}
+        onSync={() => {
+          void syncTaps().catch((syncError) => {
+            setError(syncError instanceof Error ? syncError.message : String(syncError));
+          });
+        }}
+        onRemoveTap={(tapToRemove) => {
+          void removeTap(tapToRemove).catch((removeError) => {
+            setError(removeError instanceof Error ? removeError.message : String(removeError));
+          });
+        }}
+        onPinTap={(tapToPin) => {
+          setNotice(null);
+          void pinTapToCurrentVersion(tapToPin)
+            .then(() =>
+              setNotice(`Pinned ${tapToPin.url} to ${getTapHeadCommit(tapToPin)?.slice(0, 12) ?? 'current commit'}.`)
+            )
+            .catch((pinError) => {
+              setError(pinError instanceof Error ? pinError.message : String(pinError));
+            });
+        }}
+        onUnpinTap={(tapToUnpin) => {
+          setNotice(null);
+          void unpinTap(tapToUnpin)
+            .then(() => setNotice(`Unpinned ${tapToUnpin.url}; next sync will track branch head.`))
+            .catch((unpinError) => {
+              setError(unpinError instanceof Error ? unpinError.message : String(unpinError));
+            });
+        }}
+        getTapHeadCommit={getTapHeadCommit}
+      />
 
-        <div className="crosshook-community-browser__toolbar">
-          <div className="crosshook-community-browser__field">
-            <label className="crosshook-label" htmlFor="tap-url">
-              Tap URL
-            </label>
-            <input
-              id="tap-url"
-              className="crosshook-input"
-              value={tapUrl}
-              onChange={(event) => setTapUrl(event.target.value)}
-              placeholder="https://github.com/example/community-profiles.git"
-            />
-          </div>
-          <div className="crosshook-community-browser__field">
-            <label className="crosshook-label" htmlFor="tap-branch">
-              Branch
-            </label>
-            <input
-              id="tap-branch"
-              className="crosshook-input"
-              value={tapBranch}
-              onChange={(event) => setTapBranch(event.target.value)}
-              placeholder="main"
-            />
-          </div>
-          <button
-            type="button"
-            className="crosshook-button"
-            onClick={() => {
-              void handleAddTap();
-            }}
-            disabled={loading || syncing || tapUrl.trim().length === 0}
-          >
-            Add Tap
-          </button>
-        </div>
+      <CommunityProfilesSection
+        visibleEntries={visibleEntries}
+        totalEntries={index.entries.length}
+        diagnostics={index.diagnostics}
+        query={query}
+        ratingFilter={ratingFilter}
+        loading={loading}
+        importing={importing}
+        notice={notice}
+        error={error}
+        importedProfileNames={importedProfileNames}
+        onQueryChange={setQuery}
+        onRatingFilterChange={setRatingFilter}
+        onImportFromFile={() => void handleImportFromFile()}
+        onImportEntry={(entry) => void handleImportEntry(entry)}
+      />
 
-        {taps.length > 0 ? (
-          <div className="crosshook-community-browser__tap-list">
-            {taps.map((tap) => (
-              <TapChip
-                key={tapSubscriptionStableKey(tap)}
-                tap={tap}
-                headCommit={getTapHeadCommit(tap)}
-                busy={loading || syncing}
-                onRemove={(tapToRemove) => {
-                  void removeTap(tapToRemove).catch((removeError) => {
-                    setError(removeError instanceof Error ? removeError.message : String(removeError));
-                  });
-                }}
-                onPin={(tapToPin) => {
-                  setNotice(null);
-                  void pinTapToCurrentVersion(tapToPin)
-                    .then(() =>
-                      setNotice(
-                        `Pinned ${tapToPin.url} to ${getTapHeadCommit(tapToPin)?.slice(0, 12) ?? 'current commit'}.`
-                      )
-                    )
-                    .catch((pinError) => {
-                      setError(pinError instanceof Error ? pinError.message : String(pinError));
-                    });
-                }}
-                onUnpin={(tapToUnpin) => {
-                  setNotice(null);
-                  void unpinTap(tapToUnpin)
-                    .then(() => setNotice(`Unpinned ${tapToUnpin.url}; next sync will track branch head.`))
-                    .catch((unpinError) => {
-                      setError(unpinError instanceof Error ? unpinError.message : String(unpinError));
-                    });
-                }}
-              />
-            ))}
-          </div>
-        ) : (
-          <p className="crosshook-muted crosshook-community-browser__helper">
-            Add a tap URL to populate the community browser.
-          </p>
-        )}
-      </CollapsibleSection>
-
-      <CollapsibleSection
-        title="Community Profiles"
-        className="crosshook-panel crosshook-community-browser__panel"
-        meta={
-          <span>
-            {visibleEntries.length} of {index.entries.length} profiles
-          </span>
-        }
-      >
-        <div className="crosshook-community-browser__toolbar">
-          <div className="crosshook-community-browser__field">
-            <label className="crosshook-label" htmlFor="community-search">
-              Search profiles
-            </label>
-            <input
-              id="community-search"
-              className="crosshook-input"
-              value={query}
-              onChange={(event) => setQuery(event.target.value)}
-              placeholder="Search game, trainer, author, tag..."
-            />
-          </div>
-          <div className="crosshook-community-browser__field">
-            <label className="crosshook-label" htmlFor="compatibility-filter">
-              Compatibility
-            </label>
-            <ThemedSelect
-              id="compatibility-filter"
-              value={ratingFilter}
-              onValueChange={(val) => setRatingFilter(val as 'all' | CommunityCompatibilityRating)}
-              options={[
-                { value: 'all', label: 'All ratings' },
-                ...ratingOrder.map((rating) => ({ value: rating, label: ratingLabel[rating] })),
-              ]}
-            />
-          </div>
-          <button
-            type="button"
-            className="crosshook-button crosshook-button--secondary"
-            onClick={() => void handleImportFromFile()}
-          >
-            Import JSON
-          </button>
-        </div>
-
-        {notice ? <p className="crosshook-success crosshook-community-browser__helper">{notice}</p> : null}
-        {error ? <p className="crosshook-community-browser__error">{error}</p> : null}
-        {index.diagnostics.length > 0 ? (
-          <div className="crosshook-community-browser__diagnostics">
-            {index.diagnostics.map((diagnostic) => (
-              <p key={diagnostic} className="crosshook-community-browser__diagnostic">
-                {diagnostic}
-              </p>
-            ))}
-          </div>
-        ) : null}
-
-        {loading ? (
-          <p className="crosshook-muted crosshook-community-browser__helper">Loading community profiles...</p>
-        ) : visibleEntries.length === 0 ? (
-          <p className="crosshook-community-browser__empty">
-            No community profiles matched the current search. Sync a tap or widen the filter.
-          </p>
-        ) : (
-          <div className="crosshook-community-browser__profile-grid">
-            {visibleEntries.map((entry) => {
-              const importedProfileName = deriveCommunityImportProfileName(entry);
-              const isImported = importedProfileNames.has(importedProfileName);
-              return (
-                <article
-                  key={`${entry.tap_url}::${entry.relative_path}`}
-                  className="crosshook-community-browser__profile-card"
-                >
-                  <div className="crosshook-community-browser__profile-header">
-                    <div className="crosshook-community-browser__profile-title">
-                      <h3 className="crosshook-community-browser__profile-name">
-                        {entry.manifest.metadata.game_name || 'Untitled profile'}
-                      </h3>
-                      <div className="crosshook-muted crosshook-community-browser__profile-author">
-                        {entry.manifest.metadata.author || 'Unknown author'}
-                      </div>
-                    </div>
-                    <CompatibilityBadge rating={entry.manifest.metadata.compatibility_rating} />
-                  </div>
-
-                  <div className="crosshook-community-browser__meta-grid">
-                    <div className="crosshook-muted crosshook-community-browser__meta-line">
-                      Trainer: {entry.manifest.metadata.trainer_name || 'Unknown'}{' '}
-                      {entry.manifest.metadata.trainer_version ? `(${entry.manifest.metadata.trainer_version})` : ''}
-                    </div>
-                    <div className="crosshook-muted crosshook-community-browser__meta-line">
-                      Proton: {entry.manifest.metadata.proton_version || 'Unknown'}
-                    </div>
-                    <div className="crosshook-muted crosshook-community-browser__meta-line">
-                      Game version: {entry.manifest.metadata.game_version || 'Unknown'}
-                    </div>
-                    <p className="crosshook-heading-copy crosshook-community-browser__description">
-                      {entry.manifest.metadata.description || 'No description provided.'}
-                    </p>
-                  </div>
-
-                  <div className="crosshook-community-browser__chip-row">
-                    {entry.manifest.metadata.platform_tags.length > 0 ? (
-                      entry.manifest.metadata.platform_tags.map((tag) => (
-                        <span key={tag} className="crosshook-community-browser__platform-tag">
-                          {tag}
-                        </span>
-                      ))
-                    ) : (
-                      <span className="crosshook-muted crosshook-community-browser__platform-tag crosshook-community-browser__platform-tag--empty">
-                        No platform tags
-                      </span>
-                    )}
-                  </div>
-
-                  <div className="crosshook-muted crosshook-community-browser__source">Source: {entry.tap_url}</div>
-
-                  <div className="crosshook-community-browser__button-row">
-                    {isImported ? (
-                      <span className="crosshook-community-browser__imported-badge">Imported</span>
-                    ) : (
-                      <button
-                        type="button"
-                        className="crosshook-button"
-                        onClick={() => {
-                          void handleImportEntry(entry);
-                        }}
-                        disabled={importing}
-                      >
-                        {importing ? 'Importing...' : 'Import'}
-                      </button>
-                    )}
-                  </div>
-                </article>
-              );
-            })}
-          </div>
-        )}
-      </CollapsibleSection>
       <CommunityImportWizardModal
         open={importDraft !== null}
         draft={importDraft}
